Add trailing lines and refactor several files
