--- api/help.php
@@ -1,4 +1,0 @@
-<?php
-require_once("/home/cabox/workspace/constants.php");
-
-echo "API v0.1";
