+++ database_constants.php
@@ -1,0 +1,8 @@
+<?php
+require_once("/home/cabox/workspace/constants.php");
+
+//Database constants
+define("DB_HOST", "localhost");
+define("DB_USER", "admin");
+define("DB_PASSWORD", "password");
+define("DB_NAME", "php_api");
